For issue #5: No Text Displayed in the testbed Window
Not finding the external file for the text font data has been an issue for Box2D since before I
forked it. Users run into this because the Testbed often isn't in with a working directory from
where the data file can be found. Rather than relying on software installation (which isn't always
run) putting things in the needed places, I've taken the approach of compiling the font data
directly into the Testbed application. This should provide a more reliable solution at the expense
of increasing the size of the Testbed executable file.