Changes to increase unit test coverage.
This incidentally adds reverse iteration and operator==,!=,<,<=,>,>= support to the Vector template code.
This in turn supports the use of std::set<Length2D>() to accumulate the number of vertices generated by the CreateSquareEnclosingBody function to see that it's indeed 4 as expected regardless of the shapes used to implement it.
Also increases the default vertex radius for all Shape subclasses to twice the default linear slop so as to avoid possibility of creating shapes like ChainShape that didn't have enough skin for collision handling with say polygon shapes.