Transforms Vector <, <=, >, >= support to more generic lexicographical function objects operating over containers.
Vectors don't inherently have a natural singular concept of these comparisons anyway. Arguably it's more natural for them to use these comparisons based on magnitudes. But that fails to provide regularity with respect to equality - i.e. !(A < B) && !(B < A) should mean that A == B. In the case of using magnitudes, it doesn't.