Removes Erin Catto http://www.box2d.org from copyrights of these...
While I'm not a lawyer, I believe that only the Box2D ideas of shape and
joint remain in these files; not any expression of those ideas. In this
regard, I think these are more akin to transformative works than
derivative works so I'm updating the copyright and hopefully doing so
appropriately. Here's a reference to guidance I've used for this:
https://www.nolo.com/legal-encyclopedia/fair-use-what-transformative.html