Moves WorldJoint test from Joint.cpp to WorldJoint.cpp...
Besides this being more consistent with the intention, this also made it
possible to remove World related includes from Joint.cpp. This removes all
the other declarations and definitions those files bring with them. That
improved the quality of compiler messages for example when confirming
that Joint{} == Joint{} wouldn't compile.