Fixes Doxygen docs...
JointConf is certainly not abstract. So that was just wrong.

That JointConf is not meant to be used directly isn't really true
either. At least not so much anymore. Now it's just meant to provide
data as a concrete regular type for whatever usages it may have. In this
case those usages are for specific joint configurations. Arguably if
there's an issue, it's more one of whether specific joint configuration
classes like WheelJointConf should inherit from this class, contain it,
or essentially copy it instead. These alternatives all have pros and
cons. So saying this class isn't meant to be used without deriving from
it is biased in favor of inheritance without that being my preference
now.