Extracts type calculation.
This type gets rechecked within an inner loop; a check which could conceivably be hoisted out from that loop by instead using a loop for each type.
While that would have an undesirable effect of making for three times as much code, that might have the more desirable effect of speeding up the code.
I may revisit this someday. In the meantime, this extraction into a function should help me recognize this possibility in the future.
This extraction is also appealing for helping recognize the meaning behind the Type enum value of the structure.