Changes body angle to be from transformation & adds sweep info section
Both the body's transformation and its sweep can be used to tell the
body's angle. The angle from the transformation is natually normalized
however and is always with -Pi to +Pi. Meanwhile, the angle from the
sweep pos1 is an accumulated value that's not guaranteed to be limited
to that range. Normalizing the sweep pos1 angle though should give very
closely the same angle as from the transformation. By showing both the
angular position from the body's transformation and the angular
positions from the sweep, more sense can be made out of these values.